+++ src/init.c
@@ -125,12 +125,12 @@ extern SEXP vctrs_df_list(SEXP, SEXP, SEXP);
 extern SEXP vctrs_identify_runs(SEXP);
 extern SEXP vctrs_locate_runs(SEXP, SEXP);
 extern SEXP vctrs_detect_runs(SEXP, SEXP);
-extern SEXP vctrs_df_slice_complete(SEXP);
-extern SEXP vctrs_df_locate_complete(SEXP);
-extern SEXP vctrs_df_detect_complete(SEXP);
+extern SEXP vctrs_slice_complete(SEXP);
+extern SEXP vctrs_locate_complete(SEXP);
+extern SEXP vctrs_detect_complete(SEXP);
 extern SEXP vctrs_normalize_encoding(SEXP);
-extern SEXP vctrs_order(SEXP, SEXP, SEXP);
-extern SEXP vctrs_order_locs(SEXP, SEXP, SEXP);
+extern SEXP vctrs_order(SEXP, SEXP, SEXP, SEXP);
+extern SEXP vctrs_order_locs(SEXP, SEXP, SEXP, SEXP);
 extern SEXP vctrs_unrep(SEXP);
 extern SEXP vctrs_fill_missing(SEXP, SEXP, SEXP);
 extern SEXP vctrs_chr_paste_prefix(SEXP, SEXP, SEXP);
@@ -276,12 +276,12 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_identify_runs",              (DL_FUNC) &vctrs_identify_runs, 1},
   {"vctrs_locate_runs",                (DL_FUNC) &vctrs_locate_runs, 2},
   {"vctrs_detect_runs",                (DL_FUNC) &vctrs_detect_runs, 2},
-  {"vctrs_df_slice_complete",          (DL_FUNC) &vctrs_df_slice_complete, 1},
-  {"vctrs_df_locate_complete",         (DL_FUNC) &vctrs_df_locate_complete, 1},
-  {"vctrs_df_detect_complete",         (DL_FUNC) &vctrs_df_detect_complete, 1},
+  {"vctrs_slice_complete",             (DL_FUNC) &vctrs_slice_complete, 1},
+  {"vctrs_locate_complete",            (DL_FUNC) &vctrs_locate_complete, 1},
+  {"vctrs_detect_complete",            (DL_FUNC) &vctrs_detect_complete, 1},
   {"vctrs_normalize_encoding",         (DL_FUNC) &vctrs_normalize_encoding, 1},
-  {"vctrs_order",                      (DL_FUNC) &vctrs_order, 3},
-  {"vctrs_order_locs",                 (DL_FUNC) &vctrs_order_locs, 3},
+  {"vctrs_order",                      (DL_FUNC) &vctrs_order, 4},
+  {"vctrs_order_locs",                 (DL_FUNC) &vctrs_order_locs, 4},
   {"vctrs_unrep",                      (DL_FUNC) &vctrs_unrep, 1},
   {"vctrs_fill_missing",               (DL_FUNC) &vctrs_fill_missing, 3},
   {"vctrs_chr_paste_prefix",           (DL_FUNC) &vctrs_chr_paste_prefix, 3},