+++ src/order-radix.c
@@ -14,10 +14,12 @@
 #include "utils.h"
 #include "lazy.h"
 #include "type-data-frame.h"
+#include "translate.h"
 #include "order-radix.h"
 #include "order-groups.h"
 #include "order-truelength.h"
 #include "order-sortedness.h"
+#include "order-transform.h"
 
 // -----------------------------------------------------------------------------
 
@@ -118,8 +120,9 @@
  * - `chr_order_insertion()` - Used when `x` is small.
  *
  * - `chr_order_radix()` - Same principle as integer/double ordering, but
- *   we iterate 1 character at a time. We assume a C locale here, and any
- *   non-ASCII and non-UTF8 strings are translated to UTF8.
+ *   we iterate 1 character at a time. We assume a C locale here. Any non-ASCII
+ *   and non-UTF-8 strings are translated up front by
+ *   `vec_normalize_encoding()`.
  *
  * -----------------------------------------------------------------------------
  * Logicals
@@ -181,37 +184,41 @@
 
 // -----------------------------------------------------------------------------
 
-static SEXP vec_order(SEXP x, SEXP decreasing, SEXP na_last);
+static SEXP vec_order(SEXP x, SEXP decreasing, SEXP na_last, SEXP chr_transform);
 
 // [[ register() ]]
-SEXP vctrs_order(SEXP x, SEXP direction, SEXP na_value) {
+SEXP vctrs_order(SEXP x, SEXP direction, SEXP na_value, SEXP chr_transform) {
   SEXP decreasing = PROTECT(parse_direction(direction));
   SEXP na_last = PROTECT(parse_na_value(na_value));
 
-  SEXP out = vec_order(x, decreasing, na_last);
+  SEXP out = vec_order(x, decreasing, na_last, chr_transform);
 
   UNPROTECT(2);
   return out;
 }
 
 
-static SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, bool locations);
+static SEXP vec_order_impl(SEXP x,
+                           SEXP decreasing,
+                           SEXP na_last,
+                           SEXP chr_transform,
+                           bool locations);
 
 static
-SEXP vec_order(SEXP x, SEXP decreasing, SEXP na_last) {
-  return vec_order_impl(x, decreasing, na_last, false);
+SEXP vec_order(SEXP x, SEXP decreasing, SEXP na_last, SEXP chr_transform) {
+  return vec_order_impl(x, decreasing, na_last, chr_transform, false);
 }
 
 // -----------------------------------------------------------------------------
 
-static SEXP vec_order_locs(SEXP x, SEXP decreasing, SEXP na_last);
+static SEXP vec_order_locs(SEXP x, SEXP decreasing, SEXP na_last, SEXP chr_transform);
 
 // [[ register() ]]
-SEXP vctrs_order_locs(SEXP x, SEXP direction, SEXP na_value) {
+SEXP vctrs_order_locs(SEXP x, SEXP direction, SEXP na_value, SEXP chr_transform) {
   SEXP decreasing = PROTECT(parse_direction(direction));
   SEXP na_last = PROTECT(parse_na_value(na_value));
 
-  SEXP out = vec_order_locs(x, decreasing, na_last);
+  SEXP out = vec_order_locs(x, decreasing, na_last, chr_transform);
 
   UNPROTECT(2);
   return out;
@@ -219,8 +226,8 @@ SEXP vctrs_order_locs(SEXP x, SEXP direction, SEXP na_value) {
 
 
 static
-SEXP vec_order_locs(SEXP x, SEXP decreasing, SEXP na_last) {
-  return vec_order_impl(x, decreasing, na_last, true);
+SEXP vec_order_locs(SEXP x, SEXP decreasing, SEXP na_last, SEXP chr_transform) {
+  return vec_order_impl(x, decreasing, na_last, chr_transform, true);
 }
 
 // -----------------------------------------------------------------------------
@@ -246,7 +253,6 @@ static void vec_order_switch(SEXP x,
                              struct lazy_raw* p_lazy_bytes,
                              struct lazy_raw* p_lazy_counts,
                              struct group_infos* p_group_infos,
-                             struct lazy_chr* p_lazy_x_reencoded,
                              struct truelength_info* p_truelength_info);
 
 /*
@@ -257,7 +263,7 @@ static void vec_order_switch(SEXP x,
  * the locations in `x` corresponding to each key.
  */
 static
-SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, bool locations) {
+SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, SEXP chr_transform, bool locations) {
   int n_prot = 0;
   int* p_n_prot = &n_prot;
 
@@ -267,6 +273,8 @@ SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, bool locations) {
   na_last = VECTOR_ELT(args, 1);
 
   SEXP proxy = PROTECT_N(vec_proxy_order(x), p_n_prot);
+  proxy = PROTECT_N(vec_normalize_encoding(proxy), p_n_prot);
+  proxy = PROTECT_N(proxy_chr_transform(proxy, chr_transform), p_n_prot);
 
   r_ssize size = vec_size(proxy);
   const enum vctrs_type type = vec_proxy_typeof(proxy);
@@ -325,9 +333,6 @@ SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, bool locations) {
   struct truelength_info* p_truelength_info = new_truelength_info(size);
   PROTECT_TRUELENGTH_INFO(p_truelength_info, p_n_prot);
 
-  struct lazy_chr* p_lazy_x_reencoded = new_lazy_chr(size);
-  PROTECT_LAZY_VEC(p_lazy_x_reencoded, p_n_prot);
-
   struct order* p_order = new_order(size);
   PROTECT_ORDER(p_order, p_n_prot);
 
@@ -344,7 +349,6 @@ SEXP vec_order_impl(SEXP x, SEXP decreasing, SEXP na_last, bool locations) {
     p_lazy_bytes,
     p_lazy_counts,
     p_group_infos,
-    p_lazy_x_reencoded,
     p_truelength_info
   );
 
@@ -430,7 +434,6 @@ static void df_order(SEXP x,
                      struct lazy_raw* p_lazy_bytes,
                      struct lazy_raw* p_lazy_counts,
                      struct group_infos* p_group_infos,
-                     struct lazy_chr* p_lazy_x_reencoded,
                      struct truelength_info* p_truelength_info);
 
 static void vec_order_base_switch(SEXP x,
@@ -445,7 +448,6 @@ static void vec_order_base_switch(SEXP x,
                                   struct lazy_raw* p_lazy_bytes,
                                   struct lazy_raw* p_lazy_counts,
                                   struct group_infos* p_group_infos,
-                                  struct lazy_chr* p_lazy_x_reencoded,
                                   struct truelength_info* p_truelength_info);
 
 static
@@ -461,7 +463,6 @@ void vec_order_switch(SEXP x,
                       struct lazy_raw* p_lazy_bytes,
                       struct lazy_raw* p_lazy_counts,
                       struct group_infos* p_group_infos,
-                      struct lazy_chr* p_lazy_x_reencoded,
                       struct truelength_info* p_truelength_info) {
   if (type == vctrs_type_dataframe) {
     df_order(
@@ -476,7 +477,6 @@ void vec_order_switch(SEXP x,
       p_lazy_bytes,
       p_lazy_counts,
       p_group_infos,
-      p_lazy_x_reencoded,
       p_truelength_info
     );
 
@@ -515,7 +515,6 @@ void vec_order_switch(SEXP x,
     p_lazy_bytes,
     p_lazy_counts,
     p_group_infos,
-    p_lazy_x_reencoded,
     p_truelength_info
   );
 }
@@ -581,7 +580,6 @@ static void chr_order(SEXP x,
                       struct lazy_raw* p_lazy_bytes,
                       struct lazy_raw* p_lazy_counts,
                       struct group_infos* p_group_infos,
-                      struct lazy_chr* p_lazy_x_reencoded,
                       struct truelength_info* p_truelength_info);
 
 // Used on bare vectors and the first column of data frame `x`s
@@ -598,7 +596,6 @@ void vec_order_base_switch(SEXP x,
                            struct lazy_raw* p_lazy_bytes,
                            struct lazy_raw* p_lazy_counts,
                            struct group_infos* p_group_infos,
-                           struct lazy_chr* p_lazy_x_reencoded,
                            struct truelength_info* p_truelength_info) {
   switch (type) {
   case vctrs_type_integer: {
@@ -682,7 +679,6 @@ void vec_order_base_switch(SEXP x,
       p_lazy_bytes,
       p_lazy_counts,
       p_group_infos,
-      p_lazy_x_reencoded,
       p_truelength_info
     );
 
@@ -2590,7 +2586,6 @@ static void chr_mark_sorted_uniques(const SEXP* p_x,
                                     r_ssize size,
                                     struct lazy_raw* p_lazy_x_aux,
                                     struct lazy_raw* p_lazy_bytes,
-                                    struct lazy_chr* p_lazy_x_reencoded,
                                     struct truelength_info* p_truelength_info);
 
 static inline void chr_extract_ordering(const SEXP* p_x, r_ssize size, int* p_x_aux);
@@ -2609,12 +2604,10 @@ static void chr_order_radix(const r_ssize size,
  * `chr_order_chunk()` assumes `p_x` is modifiable by reference. It also
  * assumes that `chr_mark_sorted_uniques()` has already been called. For data
  * frame columns where `chr_order_chunk()` is called on each group chunk,
- * `chr_mark_sorted_uniques()` is only called once on the entire column. It
- * also assumes that `p_x` has already been re-encoded as UTF-8 if required.
+ * `chr_mark_sorted_uniques()` is only called once on the entire column.
  *
  * `chr_order()` assumes `x` is user input which cannot be modified.
- * It copies `x` into another SEXP that can be modified directly and re-encodes
- * as UTF-8 if required.
+ * It copies `x` into another SEXP that can be modified directly.
  *
  * `chr_order_chunk()` essentially calls `int_order_chunk()`, however we can't
  * call it directly because we don't have access to all the required arguments.
@@ -2641,17 +2634,11 @@ void chr_order_chunk(bool decreasing,
                      struct group_infos* p_group_infos) {
   void* p_x_chunk = p_lazy_x_chunk->p_data;
 
-  // Don't check encoding on `p_x_chunk` data. In `df_order()`, we already
-  // ran `chr_mark_sorted_uniques()` which told `df_order()` whether or not
-  // to re-encode as it created `p_x_chunk`
-  bool check_encoding = false;
-
   const enum vctrs_sortedness sortedness = chr_sortedness(
     p_x_chunk,
     size,
     decreasing,
     na_last,
-    check_encoding,
     p_group_infos
   );
 
@@ -2699,7 +2686,6 @@ struct chr_order_info {
   struct lazy_raw* p_lazy_bytes;
   struct lazy_raw* p_lazy_counts;
   struct group_infos* p_group_infos;
-  struct lazy_chr* p_lazy_x_reencoded;
   struct truelength_info* p_truelength_info;
 };
 
@@ -2728,7 +2714,6 @@ void chr_order(SEXP x,
                struct lazy_raw* p_lazy_bytes,
                struct lazy_raw* p_lazy_counts,
                struct group_infos* p_group_infos,
-               struct lazy_chr* p_lazy_x_reencoded,
                struct truelength_info* p_truelength_info) {
   struct chr_order_info info = {
     .x = x,
@@ -2742,7 +2727,6 @@ void chr_order(SEXP x,
     .p_lazy_bytes = p_lazy_bytes,
     .p_lazy_counts = p_lazy_counts,
     .p_group_infos = p_group_infos,
-    .p_lazy_x_reencoded = p_lazy_x_reencoded,
     .p_truelength_info = p_truelength_info
   };
 
@@ -2769,7 +2753,6 @@ static void chr_order_internal(SEXP x,
                                struct lazy_raw* p_lazy_bytes,
                                struct lazy_raw* p_lazy_counts,
                                struct group_infos* p_group_infos,
-                               struct lazy_chr* p_lazy_x_reencoded,
                                struct truelength_info* p_truelength_info);
 
 static
@@ -2788,7 +2771,6 @@ SEXP chr_order_exec(void* p_data) {
     p_info->p_lazy_bytes,
     p_info->p_lazy_counts,
     p_info->p_group_infos,
-    p_info->p_lazy_x_reencoded,
     p_info->p_truelength_info
   );
 
@@ -2813,19 +2795,14 @@ void chr_order_internal(SEXP x,
                         struct lazy_raw* p_lazy_bytes,
                         struct lazy_raw* p_lazy_counts,
                         struct group_infos* p_group_infos,
-                        struct lazy_chr* p_lazy_x_reencoded,
                         struct truelength_info* p_truelength_info) {
   const SEXP* p_x = STRING_PTR_RO(x);
-
-  // Check encodings when determining sortedness of user input
-  bool check_encoding = true;
 
   const enum vctrs_sortedness sortedness = chr_sortedness(
     p_x,
     size,
     decreasing,
     na_last,
-    check_encoding,
     p_group_infos
   );
 
@@ -2839,22 +2816,14 @@ void chr_order_internal(SEXP x,
 
   // Sort unique strings and mark their truelengths with ordering.
   // Use `p_lazy_x_chunk` as auxiliary memory for `chr_order_radix()` so we
-  // hopefully don't have to also allocate `p_lazy_x_aux`. If re-encoding
-  // is required, it stores the results in `p_lazy_x_reencoded`.
+  // hopefully don't have to also allocate `p_lazy_x_aux`.
   chr_mark_sorted_uniques(
     p_x,
     size,
     p_lazy_x_chunk,
     p_lazy_bytes,
-    p_lazy_x_reencoded,
     p_truelength_info
   );
-
-  // If we re-encoded, then the vector to extract the ordering from is in
-  // `p_lazy_x_reencoded`.
-  if (p_truelength_info->reencode) {
-    p_x = p_lazy_x_reencoded->p_data;
-  }
 
   void* p_x_chunk = init_lazy_raw(p_lazy_x_chunk);
 
@@ -2929,12 +2898,6 @@ static void chr_mark_uniques(const SEXP* p_x,
  * through `p_x` and just pluck off the TRUELENGTH value, which will be an
  * integer proxy for the value's ordering.
  *
- * We optimize heavily for the ASCII / UTF-8 case by checking the encodings of
- * only the uniques. If any uniques need re-encoding, we recompute the unique
- * strings again on the entire vector, this time with reencoding. This gives
- * a nice speed boost for the most common case of all ASCII/UTF-8 because
- * checking encodings is expensive.
- *
  * `truelength_save()` also saves the unique strings and their original
  * TRUELENGTH values so they can be reset after each column with
  * `truelength_reset()`.
@@ -2944,33 +2907,8 @@ void chr_mark_sorted_uniques(const SEXP* p_x,
                              r_ssize size,
                              struct lazy_raw* p_lazy_x_aux,
                              struct lazy_raw* p_lazy_bytes,
-                             struct lazy_chr* p_lazy_x_reencoded,
                              struct truelength_info* p_truelength_info) {
   chr_mark_uniques(p_x, size, p_truelength_info);
-
-  // Check if any uniques need reencoding
-  bool reencode = p_chr_any_reencode(
-    p_truelength_info->p_uniques,
-    p_truelength_info->size_used
-  );
-
-  // Rerun marking of unique values if any needed reencoding
-  // (some characters might translate to the same UTF-8 character)
-  if (reencode) {
-    // Reset existing uniques before rerun
-    truelength_reset(p_truelength_info);
-
-    // Initialize container for re-encoded result
-    init_lazy_chr(p_lazy_x_reencoded);
-
-    p_chr_copy_with_reencode(p_x, p_lazy_x_reencoded->data, size);
-
-    // Tell `df_order()` and `chr_order()` we re-encoded
-    p_truelength_info->reencode = true;
-
-    // Re-mark uniques on re-encoded vector
-    chr_mark_uniques(p_lazy_x_reencoded->p_data, size, p_truelength_info);
-  }
 
   r_ssize n_uniques = p_truelength_info->size_used;
 
@@ -3352,7 +3290,6 @@ struct df_order_info {
   struct lazy_raw* p_lazy_bytes;
   struct lazy_raw* p_lazy_counts;
   struct group_infos* p_group_infos;
-  struct lazy_chr* p_lazy_x_reencoded;
   struct truelength_info* p_truelength_info;
 };
 
@@ -3393,7 +3330,6 @@ void df_order(SEXP x,
               struct lazy_raw* p_lazy_bytes,
               struct lazy_raw* p_lazy_counts,
               struct group_infos* p_group_infos,
-              struct lazy_chr* p_lazy_x_reencoded,
               struct truelength_info* p_truelength_info) {
   struct df_order_info info = {
     .x = x,
@@ -3407,7 +3343,6 @@ void df_order(SEXP x,
     .p_lazy_bytes = p_lazy_bytes,
     .p_lazy_counts = p_lazy_counts,
     .p_group_infos = p_group_infos,
-    .p_lazy_x_reencoded = p_lazy_x_reencoded,
     .p_truelength_info = p_truelength_info
   };
 
@@ -3434,7 +3369,6 @@ static void df_order_internal(SEXP x,
                               struct lazy_raw* p_lazy_bytes,
                               struct lazy_raw* p_lazy_counts,
                               struct group_infos* p_group_infos,
-                              struct lazy_chr* p_lazy_x_reencoded,
                               struct truelength_info* p_truelength_info);
 
 static
@@ -3453,7 +3387,6 @@ SEXP df_order_exec(void* p_data) {
     p_info->p_lazy_bytes,
     p_info->p_lazy_counts,
     p_info->p_group_infos,
-    p_info->p_lazy_x_reencoded,
     p_info->p_truelength_info
   );
 
@@ -3526,7 +3459,6 @@ void df_order_internal(SEXP x,
                        struct lazy_raw* p_lazy_bytes,
                        struct lazy_raw* p_lazy_counts,
                        struct group_infos* p_group_infos,
-                       struct lazy_chr* p_lazy_x_reencoded,
                        struct truelength_info* p_truelength_info) {
   r_ssize n_cols = r_length(x);
 
@@ -3590,7 +3522,6 @@ void df_order_internal(SEXP x,
     p_lazy_bytes,
     p_lazy_counts,
     p_group_infos,
-    p_lazy_x_reencoded,
     p_truelength_info
   );
 
@@ -3601,7 +3532,14 @@ void df_order_internal(SEXP x,
 
   // Iterate over remaining columns by group chunk
   for (r_ssize i = 1; i < n_cols; ++i) {
-    col = VECTOR_ELT(x, i);
+    // Get the number of group chunks from previous column group info
+    struct group_info* p_group_info_pre = groups_current(p_group_infos);
+    r_ssize n_groups = p_group_info_pre->n_groups;
+
+    // If there were no ties, we are completely done
+    if (n_groups == size) {
+      break;
+    }
 
     if (!recycle_decreasing) {
       col_decreasing = p_decreasing[i];
@@ -3616,15 +3554,7 @@ void df_order_internal(SEXP x,
     // processed at least one column.
     int* p_o_col = p_order->p_data;
 
-    // Get the number of group chunks from previous column group info
-    struct group_info* p_group_info_pre = groups_current(p_group_infos);
-    r_ssize n_groups = p_group_info_pre->n_groups;
-
-    // If there were no ties, we are completely done
-    if (n_groups == size) {
-      break;
-    }
-
+    col = VECTOR_ELT(x, i);
     type = vec_proxy_typeof(col);
 
     // If we are on the rerun pass, flip this back off so the
@@ -3633,7 +3563,7 @@ void df_order_internal(SEXP x,
       rerun_complex = rerun_complex ? false : true;
     }
 
-    // Pre sort unique characters once for the whole column
+    // Pre-sort unique characters once for the whole column
     if (type == vctrs_type_character) {
       const SEXP* p_col = STRING_PTR_RO(col);
 
@@ -3642,15 +3572,8 @@ void df_order_internal(SEXP x,
         size,
         p_lazy_x_aux,
         p_lazy_bytes,
-        p_lazy_x_reencoded,
         p_truelength_info
       );
-
-      // If re-encoding was required, the re-encoded column is stored
-      // in `p_lazy_x_reencoded`.
-      if (p_truelength_info->reencode) {
-        col = p_lazy_x_reencoded->data;
-      }
     }
 
     // Turn off group tracking if: